chore: move InMemoryWordProvider to its own file
