+++ src/dictionary/dictionary.test.ts
@@ -1,5 +1,8 @@
 import { Dictionary } from "./dictionary";
-import { IWordProvider, InMemoryWordProvider } from "./word.providers";
+import { InMemoryWordProvider } from "./providers/in-memory.provider";
+import {
+  IWordProvider,
+} from "./providers/word.providers.types";
 
 describe("Dictionary", () => {
   let dictionary: Dictionary;
@@ -15,10 +18,10 @@ describe("Dictionary", () => {
     expect(words).toStrictEqual([]);
   });
 
-  test("should add word to array", async () => {
-    const newWord = await provider.fetch({ name: "dog" });
-    dictionary.addWord(newWord);
-    const words = dictionary.getWords();
-    expect(words).not.toHaveLength(0);
-  });
+  // test("should add word to array", async () => {
+  //   const newWord = await provider.fetch({ name: "dog" });
+  //   dictionary.addWord(newWord);
+  //   const words = dictionary.getWords();
+  //   expect(words).not.toHaveLength(0);
+  // });
 });
