+++ src/dictionary/providers/word.providers.types.ts
@@ -1,0 +1,9 @@
+import { Word } from "../word";
+
+export type WordToFetch = {
+  name: string;
+};
+
+export interface IWordProvider {
+  fetch(word: WordToFetch): Promise<Word>;
+}
